# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WhuzPay Payment Gateway Aggregator API
  description: |
    REST API for the WhuzPay QRIS payment gateway aggregator. Orchestrates
    payment providers (Cashi in production, an in-process mock for
    merchant sandbox) behind one consistent API and dashboard.

    Generated from the actual route table in `cmd/api/main.go` and the
    handler/service/domain source — not hand-guessed. See
    [docs/payment-state-machine.md](payment-state-machine.md) for the
    payment status lifecycle these endpoints read and write.
  version: "1.0"
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
security: []
paths:
...
  /admin/payments:
    get:
      tags: [Admin - Payments]
      summary: List payments (admin view, all merchants)
      security: [{ adminBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: merchant_id, in: query, schema: { type: string, format: uuid } }
        - { name: environment, in: query, schema: { type: string, enum: [sandbox, production] } }
        - { name: date_from, in: query, schema: { type: string, format: date-time } }
        - { name: date_to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedPayments' } } } }
    post:
      tags: [Admin - Payments]
      summary: Create a payment on behalf of a merchant
      security: [{ adminBearer: [] }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePaymentRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { description: Provider error }
        '503': { description: Provider unavailable }
...
  /admin/payments/{id}:
    get:
      tags: [Admin - Payments]
      summary: Get payment detail
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AdminPaymentDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/merchants/{id}/payments:
    get:
      tags: [Admin - Merchants]
      summary: List a merchant's payments
      security: [{ adminBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/PathID' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedPayments' } } } }
...
  /merchant/payments:
    get:
      tags: [Merchant Dashboard]
      summary: List this merchant's payments
      security: [{ merchantBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: environment, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date-time } }
        - { name: date_to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedPayments' } } } }
    post:
      tags: [Merchant Dashboard]
      summary: Create a payment from the dashboard
      security: [{ merchantBearer: [] }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePaymentRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /merchant/payments/{id}:
    get:
      tags: [Merchant Dashboard]
      summary: Get one of this merchant's payments
      security: [{ merchantBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '404': { description: Not found (also returned if the payment belongs to another merchant) }
...
  /merchant/payment-links/{id}/payments:
    get:
      tags: [Merchant Payment Links]
      summary: List payments spawned by this link
      security: [{ merchantBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/PathID' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedPayments' } } } }

  # ---------------------------------------------------------------- Public
...
  /public/payment-links/{slug}/pay:
    post:
      tags: [Public]
      summary: Check out through a payment link
      description: |
        No auth. Rate limited 120 req/min per IP. Spawns a fresh one-time
        Payment linked back to this Payment Link (see
        PaymentLinkService). `amount` is only honored for open-amount
        links — for fixed links it is ignored server-side.
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/InitiateCheckoutRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  # ---------------------------------------------------------------- Merchant API (API key)
  /payments:
    post:
      tags: [Merchant API]
      summary: Create a payment (server-to-server)
      description: |
        `merchant_id` and `environment` in the request body are ignored —
        forced from the authenticated API key
        (see MerchantAPIAuthMiddleware) so a caller can never forge
        another merchant's ID.
      security: [{ apiKey: [] }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePaymentRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { description: Missing/invalid/revoked API key }
        '403': { description: Merchant inactive }
        '502': { description: Provider error }
        '503': { description: Provider unavailable }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /payments/{id}:
    get:
      tags: [Merchant API]
      summary: Get a payment (server-to-server)
      security: [{ apiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/PathID' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session token from POST /auth/admin/login.
    merchantBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Merchant dashboard session token from POST /auth/login.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Merchant API key. Accepted either as `X-API-Key: <key>` or
        `Authorization: Bearer <key>` (see MerchantAPIAuthMiddleware).
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    TooManyRequests:
      description: Rate limit exceeded (per-IP)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }

    # ---- Auth ----
...
    CreatePaymentRequest:
      type: object
      required: [amount, payment_method, description]
      properties:
        merchant_id: { type: string, format: uuid, description: Forced from context on merchant-authenticated routes }
        amount: { type: integer, minimum: 1 }
        currency: { type: string, default: IDR }
        payment_method: { type: string, enum: [qris] }
        description: { type: string }
        customer_name: { type: string, nullable: true }
        customer_email: { type: string, nullable: true }
        callback_url: { type: string, nullable: true }
        expires_in_minutes: { type: integer, minimum: 1, maximum: 1440, default: 30 }
        environment: { type: string, enum: [sandbox, production] }
        use_custom_merchant_name:
          type: boolean
          description: Requests a custom merchant display name on the QR (provider-dependent, e.g. Cashi QRIS Custom).
    PaymentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        merchant_id: { type: string, format: uuid }
        amount: { type: integer }
        currency: { type: string }
        payment_method: { type: string }
        provider_name: { type: string }
        provider_reference: { type: string, nullable: true }
        status: { type: string, enum: [pending, paid, expired, failed, cancelled] }
        description: { type: string }
        customer_name: { type: string, nullable: true }
        customer_email: { type: string, nullable: true }
        qris_data: { type: string, nullable: true }
        checkout_url: { type: string }
        environment: { type: string }
        payment_link_id: { type: string, format: uuid, nullable: true }
        expires_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    AdminPaymentItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        merchant_id: { type: string, format: uuid }
        merchant_name: { type: string }
        amount: { type: integer }
        currency: { type: string }
        payment_method: { type: string }
        provider_name: { type: string }
        provider_reference: { type: string, nullable: true }
        status: { type: string }
        description: { type: string }
        customer_name: { type: string, nullable: true }
        customer_email: { type: string, nullable: true }
        environment: { type: string }
        expires_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AdminPaymentDetail:
      allOf:
        - $ref: '#/components/schemas/AdminPaymentItem'
        - type: object
          properties:
            qris_data: { type: string, nullable: true }
            callback_url: { type: string, nullable: true }
    PaginatedPayments:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AdminPaymentItem' } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    InitiateCheckoutRequest:
      type: object
      description: amount is only honored for open-amount links.
      properties:
        amount: { type: integer }
        customer_name: { type: string, nullable: true }
        customer_email: { type: string, nullable: true }
...